Refactor: 역할 분리를 위한 디렉토리 추가 및 DTO 클래스 네이밍 수정 및 이동
- 네이밍 및 디렉토리 변경으로 인한 코드 변경 외에 변동 사항 없음
