+++ src/main/java/com/jerry/ticketing/concert/api/ConcertWebController.java
@@ -2,8 +2,8 @@ package com.jerry.ticketing.concert.api;
 
 
 import com.jerry.ticketing.concert.application.ConcertService;
-import com.jerry.ticketing.concert.application.dto.ConcertDto;
-import com.jerry.ticketing.concert.application.dto.CreateConcertDto;
+import com.jerry.ticketing.concert.application.dto.domain.ConcertDto;
+import com.jerry.ticketing.concert.application.dto.web.CreateConcertDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
