+++ src/main/java/com/jerry/ticketing/concert/infrastructure/mapper/ConcertMapper.java
@@ -1,6 +1,7 @@
-package com.jerry.ticketing.concert.domain;
+package com.jerry.ticketing.concert.infrastructure.mapper;
 
-import com.jerry.ticketing.concert.application.dto.CreateConcertDto;
+import com.jerry.ticketing.concert.application.dto.web.CreateConcertDto;
+import com.jerry.ticketing.concert.domain.Concert;
 import org.springframework.stereotype.Component;
 
 @Component
