+++ src/main/java/com/jerry/ticketing/payment/application/PaymentApplicationService.java
@@ -2,16 +2,16 @@ package com.jerry.ticketing.payment.application;
 
 
 import com.jerry.ticketing.concert.application.ConcertService;
-import com.jerry.ticketing.concert.application.dto.ConcertDto;
+import com.jerry.ticketing.concert.application.dto.domain.ConcertDto;
 import com.jerry.ticketing.global.exception.BusinessException;
 import com.jerry.ticketing.global.exception.PaymentErrorCode;
 import com.jerry.ticketing.member.application.MemberService;
-import com.jerry.ticketing.member.application.dto.MemberDto;
-import com.jerry.ticketing.payment.application.dto.CreatePaymentDto;
-import com.jerry.ticketing.payment.application.dto.PaymentDto;
+import com.jerry.ticketing.member.application.dto.domain.MemberDto;
+import com.jerry.ticketing.payment.application.dto.web.CreatePaymentDto;
+import com.jerry.ticketing.payment.application.dto.domain.PaymentDto;
 import com.jerry.ticketing.payment.infrastructure.repository.PaymentRepository;
 import com.jerry.ticketing.reservation.application.ReservationService;
-import com.jerry.ticketing.reservation.application.dto.ReservationDto;
+import com.jerry.ticketing.reservation.application.dto.domain.ReservationDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
