+++ src/main/java/com/jerry/ticketing/payment/application/PaymentService.java
@@ -4,9 +4,9 @@ import com.jerry.ticketing.global.util.PaymentOrderIdGenerator;
 import com.jerry.ticketing.payment.domain.Payment;
 import com.jerry.ticketing.payment.infrastructure.external.TossPaymentClient;
 import com.jerry.ticketing.reservation.domain.Reservation;
-import com.jerry.ticketing.payment.application.dto.ConfirmPaymentDto;
-import com.jerry.ticketing.payment.application.dto.CreatePaymentDto;
-import com.jerry.ticketing.payment.application.dto.WebhookPaymentDto;
+import com.jerry.ticketing.payment.application.dto.web.ConfirmPaymentDto;
+import com.jerry.ticketing.payment.application.dto.web.CreatePaymentDto;
+import com.jerry.ticketing.payment.application.dto.web.WebhookPaymentDto;
 import com.jerry.ticketing.global.exception.BusinessException;
 import com.jerry.ticketing.global.exception.PaymentErrorCode;
 import com.jerry.ticketing.global.exception.ReservationErrorCode;
