+++ src/main/java/com/jerry/ticketing/reservation/application/ReservationService.java
@@ -4,9 +4,9 @@ package com.jerry.ticketing.reservation.application;
 import com.jerry.ticketing.concert.domain.Concert;
 import com.jerry.ticketing.global.exception.PaymentErrorCode;
 import com.jerry.ticketing.member.domain.Member;
-import com.jerry.ticketing.reservation.application.dto.ReservationDto;
+import com.jerry.ticketing.reservation.application.dto.domain.ReservationDto;
 import com.jerry.ticketing.reservation.domain.Reservation;
-import com.jerry.ticketing.reservation.application.dto.CreateReservationDto;
+import com.jerry.ticketing.reservation.application.dto.web.CreateReservationDto;
 import com.jerry.ticketing.global.exception.BusinessException;
 import com.jerry.ticketing.global.exception.ReservationErrorCode;
 import com.jerry.ticketing.concert.infrastructure.repository.ConcertRepository;
