+++ src/main/java/com/jerry/ticketing/seat/domain/vo/ConcertSeats.java
@@ -1,4 +1,6 @@
-package com.jerry.ticketing.seat.domain;
+package com.jerry.ticketing.seat.domain.vo;
+
+import com.jerry.ticketing.seat.domain.ConcertSeat;
 
 import java.util.List;
 
@@ -9,27 +11,27 @@ public class ConcertSeats {
         this.concertSeats = concertSeats;
     }
 
-    public List<ConcertSeat> item(){
+    public List<ConcertSeat> item() {
         return this.concertSeats;
     }
 
-    public static ConcertSeats from(List<ConcertSeat> concertSeats){
+    public static ConcertSeats from(List<ConcertSeat> concertSeats) {
         return new ConcertSeats(concertSeats);
     }
 
-    public void block(Long memberId){
-        this.concertSeats.forEach(v->{
-            v.blockConcertSeat(memberId);});
+    public void block(Long memberId) {
+        this.concertSeats.forEach(v -> {
+            v.blockConcertSeat(memberId);
+        });
     }
 
-    public void available(){
+    public void available() {
         this.concertSeats.forEach(ConcertSeat::initConcertSeat);
     }
 
-    public boolean isNotSame(List<Long> ConcertSeatIds){
+    public boolean isNotSame(List<Long> ConcertSeatIds) {
         return this.concertSeats.size() != ConcertSeatIds.size();
     }
-
 
 
 }
